requirements.txt
numpy==1.20.3
paho-mqtt==1.5.1
